+++ packages/k2-alpine/src/components/StakeCards/BaseCard.tsx
@@ -1,0 +1,43 @@
+import React from 'react'
+import { SxProp } from 'dripsy'
+import { useTheme } from '../../hooks'
+import { AnimatedPressable } from '../Animated/AnimatedPressable'
+import { View } from '../Primitives'
+
+export const BaseCard = ({
+  onPress,
+  sx,
+  children
+}: {
+  onPress?: () => void
+  sx?: SxProp
+  children: React.ReactNode
+}): JSX.Element => {
+  const { theme } = useTheme()
+  const borderColor = theme.isDark ? '#FFFFFF1A' : '#0000001A'
+
+  return (
+    <AnimatedPressable style={{ flex: 1 }} onPress={onPress}>
+      <View
+        sx={{
+          borderRadius: 18,
+          borderWidth: 1,
+          borderColor,
+          backgroundColor: '$surfaceSecondary',
+          overflow: 'hidden',
+          ...sx
+        }}>
+        {children}
+      </View>
+    </AnimatedPressable>
+  )
+}
+
+export const DEFAULT_CARD_WIDTH = 200
+
+export const getCardHeight = (width: number): number => {
+  const ratio = 5 / 4
+  const minHeight = 210
+
+  return Math.floor(Math.max(width * ratio, minHeight))
+}
