+++ packages/k2-alpine/src/components/StakeCards/ProgressCard.tsx
@@ -1,0 +1,164 @@
+import React, { useEffect, useMemo, useRef } from 'react'
+import Animated, {
+  SharedValue,
+  useAnimatedProps,
+  useAnimatedStyle,
+  useDerivedValue,
+  useSharedValue,
+  withSpring,
+  withTiming
+} from 'react-native-reanimated'
+import Svg, { Path } from 'react-native-svg'
+import { DeviceMotionMeasurement } from 'expo-sensors'
+import { Text, View } from '../Primitives'
+import { useTheme } from '../../hooks'
+import { BaseCard, DEFAULT_CARD_WIDTH, getCardHeight } from './BaseCard'
+
+export const ProgressCard = ({
+  progress,
+  title,
+  width = DEFAULT_CARD_WIDTH,
+  motion
+}: ProgressCardProps): JSX.Element => {
+  const {
+    theme: { colors }
+  } = useTheme()
+
+  const phaseConstant = useMemo(() => Math.random() * 0.4 + 0.8, []) // randomize phase a bit, from 0.8 to 1.2
+
+  const phase = useSharedValue(0)
+  const height = getCardHeight(width)
+  const baseHeight = height * (1 - progress)
+
+  const lastUpdateTime = useRef(Date.now())
+
+  const waveWidth = useMemo(() => width * 2, [width])
+
+  const rotationZ = useDerivedValue(() => {
+    return withTiming(motion?.value?.rotation.gamma ?? 0, { duration: 100 })
+  })
+
+  const rotationStyle = useAnimatedStyle(() => {
+    return {
+      transform: [{ rotateZ: `${-rotationZ.value * 0.5}rad` }]
+    }
+  })
+
+  const amplitude = useSharedValue(0)
+
+  const animatedProps = useAnimatedProps(() => {
+    const A = amplitude.value
+    const φ = phase.value
+
+    const numberOfPoints = 20
+    const step = waveWidth / (numberOfPoints - 1)
+    let d = `M 0 ${height} L 0 ${baseHeight}`
+    for (let i = 0; i < numberOfPoints; i++) {
+      const x = i * step
+      const y = baseHeight + A * Math.sin((x / waveWidth) * 2 * Math.PI + φ)
+      d += ` L ${x} ${y}`
+    }
+    d += ` L ${waveWidth} ${height} Z`
+
+    return { d }
+  })
+
+  const fillColor = colors.$borderPrimary
+
+  useDerivedValue(() => {
+    if (!motion?.value) return
+
+    const { accelerationIncludingGravity } = motion.value
+    if (accelerationIncludingGravity) {
+      const accMagnitude = Math.sqrt(
+        accelerationIncludingGravity.x ** 2 +
+          accelerationIncludingGravity.y ** 2 +
+          accelerationIncludingGravity.z ** 2
+      )
+      if (accMagnitude > 10.5) {
+        amplitude.value = withTiming(accMagnitude, { duration: 300 })
+        lastUpdateTime.current = Date.now()
+      }
+    }
+  }, [motion])
+
+  // reset amplitude if no motion detected for 1 second
+  useEffect(() => {
+    if (motion) {
+      const interval = setInterval(() => {
+        const now = Date.now()
+        if (now - lastUpdateTime.current > 1000) {
+          amplitude.value = withSpring(0, { duration: 10000 })
+        }
+      }, 1000)
+      return () => clearInterval(interval)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [motion])
+
+  useEffect(() => {
+    let animationFrameId: number
+
+    const updatePhase = (): void => {
+      if (motion?.value) {
+        phase.value = (phase.value + 0.1 * phaseConstant) % (2 * Math.PI)
+      }
+      animationFrameId = requestAnimationFrame(updatePhase)
+    }
+
+    animationFrameId = requestAnimationFrame(updatePhase)
+
+    return () => {
+      cancelAnimationFrame(animationFrameId)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [phaseConstant, motion])
+
+  return (
+    <BaseCard
+      sx={{
+        paddingVertical: 20,
+        paddingHorizontal: 18,
+        width,
+        height
+      }}>
+      <View
+        style={{
+          position: 'absolute',
+          left: 0,
+          right: 0,
+          top: 0,
+          bottom: 0
+        }}>
+        <Animated.View
+          style={[
+            {
+              position: 'absolute',
+              left: -width / 2,
+              right: -width / 2,
+              top: 0,
+              bottom: -height / 2
+            },
+            rotationStyle
+          ]}>
+          <Svg width={waveWidth} height={height}>
+            <AnimatedPath animatedProps={animatedProps} fill={fillColor} />
+          </Svg>
+          <View style={{ backgroundColor: fillColor, flex: 1 }} />
+        </Animated.View>
+      </View>
+      <Text sx={{ fontFamily: 'Aeonik-Bold', fontSize: 24, lineHeight: 22 }}>
+        {title}
+      </Text>
+    </BaseCard>
+  )
+}
+
+export type ProgressCardProps = {
+  progress: number
+  title: string
+  width?: number
+  motion?: SharedValue<DeviceMotionMeasurement | undefined>
+}
+
+const AnimatedPath = Animated.createAnimatedComponent(Path)
